+++ lib/iris/fileformats/netcdf/__init__.py
@@ -1,0 +1,9 @@
+# Copyright Iris contributors
+#
+# This file is part of Iris and is released under the LGPL license.
+# See COPYING and COPYING.LESSER in the root of the repository for full
+# licensing details.
+"""
+A package for loading and saving cubes to and from netcdf files.
+
+"""
